application insight fixes
