+++ src/listeners/index.ts
@@ -1,5 +1,8 @@
+export * from './BaseListener';
 export * from './DashboardListener';
+export * from './DataListener';
 export * from './ExtensionListener';
 export * from './MediaListener';
 export * from './PagesListener';
 export * from './SettingsListener';
+export * from './TelemetryListener';
